+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the cavalier_contours crate will be documented in this file.
 
+## Unreleased
+
+### Internal
+
+- Added Criterion benchmarks for polyline area and raw and final parallel offset creation.
+
 ## 0.8.0 - 2026-08-09
 
 ### Changed 🔧